+++ truncate_gradient.h
@@ -2,7 +2,8 @@
 #define TRUNCATE_GRADIENT_H
 
 #include <math.h>  //fabs
+#include <stddef.h> //size_t
 
-void truncate_gradient(double* , int , double );
+void truncate_gradient(double* , size_t , double );
 
 #endif /* TRUNCATE_GRADIENT_H */
